Add TwitterAdapter
